refact: Use v2 CircleCI api for most commands
This uses the v2 api for everything except getting the full output of a
job for the `cci output` command, since this functionality does not
exist on v2.

This allows for a more reliable was of retrieving just the latest
pipeline run for `cci status`. Previously this would display the most
recent jobs, but they could span multiple pipelines, resulting in
confusing output.

It also changes how `cci retry` works. Previously a single job could be
retried, but this would run in "legacy" mode and would not show under
the new pipelines view of the CircleCI UI. It also would result in
confusing status reports back to services like Github. Now `cci retry`
will take a workflow name, and will retry the entire workflow.

Note that the v2 api documentation seems to indicate that specific jobs
can be selected to be retried, however this does not seem to work
currently, and the entire workflow is retried regardless of the job ids
sent. So for now `cci retry` will just take a workflow name and retry
the entire workflow.

See:
    https://circleci.com/docs/2.0/api-developers-guide/
    https://circleci.com/docs/api/v2/#operation/listPipelinesForProject
    https://circleci.com/docs/api/v2/#operation/listWorkflowsByPipelineId
    https://circleci.com/docs/api/v2/#operation/listWorkflowJobs
    https://circleci.com/docs/api/v2/#operation/rerunWorkflow